test(e2e-mitm): send CORS headers and echo the JSON-RPC id
